requirements.txt
Flask
Flask-Login
psycopg2-binary
python-dotenv
werkzeug